+++ .eslintrc.js
@@ -5,4 +5,9 @@ module.exports = {
     node: true,
   },
   extends: 'airbnb-base',
+  rules: {
+    'linebreak-style': 'off',
+    "max-classes-per-file": ["error", 2],
+    'object-curly-newline': ["error", { "minProperties": 5 }]
+  },
 };
