+++ lib/helpers/prompts.js
@@ -1,11 +1,11 @@
+/* eslint-disable no-param-reassign */
 const inquirer = require('inquirer');
-const fixture = require('./fixture.js');
-const helpers = require('./helpers.js');
+const utils = require('./utils.js');
 const { print } = require('./theme.js');
 
-async function retrieveFixture(fixtureKey, path) {
-  const componentPath = helpers.getComponentPath(path);
-  const fixtures = fixture.readFixture(componentPath);
+exports.setupFixture = async function setupFixture(fixtureKey, path) {
+  const componentPath = utils.getComponentPath(path);
+  const fixtures = utils.readFixtureFile(componentPath);
 
   if (!fixtures) {
     print.error('No fixtures found');
@@ -27,14 +27,14 @@ async function retrieveFixture(fixtureKey, path) {
   const result = fixtures[fixtureKey];
   if (!result) {
     print.error(`Fixture '${fixtureKey}' does not exist`);
-    helpers.destroyProcess();
+    utils.destroyProcess();
     throw new Error('Fixture does not exist');
   }
   return result;
-}
+};
 
-async function retrieveFunctionName(functionName, path) {
-  const component = helpers.resolveComponent(path);
+exports.setupFunction = async function setupFunction(functionName, path) {
+  const component = utils.resolveComponent(path);
 
   if (!functionName) {
     const promptMsg = 'Please select your function to execute: ';
@@ -53,11 +53,12 @@ async function retrieveFunctionName(functionName, path) {
 
   if (!result) {
     print.error(`Function '${functionName}' does not exist`);
-    helpers.destroyProcess();
+    utils.destroyProcess();
     throw new Error('Function does not exist');
   }
   return result;
-}
+};
 
-exports.retrieveFixture = retrieveFixture;
-exports.retrieveFunctionName = retrieveFunctionName;
+exports.getAction = async function getAction() {
+  const componentJSON = utils.readComponentJSON();
+};
