+++ lib/helpers/theme.js
